function: getUserInfoById

+++ src/main/java/com/baofeng/blog/mapper/UserMapper.java
@@ -9,8 +9,8 @@ import java.util.List;
 public interface UserMapper {
     int insertUser(User user);
     int updateUser(User user);
-    int deleteUser(Long id);
-    User selectById(Long id);
+    int deleteUser(int id);
+    User selectById(int id);
     
     List<User> selectByCondition(@Param("username") String username, 
                                @Param("email") String email,
